+++ sccfm_cli/commands/inventory/devices/asa/change_boot_image/__init__.py
@@ -1,0 +1,5 @@
+__all__ = ["AsaChangeBootImageCommand"]
+
+from sccfm_cli.commands.inventory.devices.asa.change_boot_image.command import (
+    AsaChangeBootImageCommand,
+)